rename main components
