+++ components/GroupsPage/Components/TeamCard.js
@@ -36,7 +36,7 @@ export default function TeamCard({ group, groupStatus }) {
 
   useEffect(() => {
     const handleTooltips = async () => {
-      const module = await import("@components/Dasboard/Tooltips");
+      const module = await import("@components/Dashboard/Components/Tooltips");
       if (module && module.handleTooltips) {
         module.handleTooltips();
       }
